added AddText component
