+++ src/component/AddText.jsx
@@ -1,0 +1,34 @@
+import { useState } from "react";
+import NewText from "../component/NewText";
+
+function AddText() {
+  const [showTextSection, setShowTextSection] = useState(false);
+
+  const handleAddTextClick = () => {
+    setShowTextSection(true);
+  };
+
+  const handleCloseTextClick = () => {
+    setShowTextSection(false);
+  };
+
+  return (
+    <div>
+      {!showTextSection && (
+        <button className="button" onClick={handleAddTextClick}>
+          <p>Add a new Text Entry for today</p>
+        </button>
+      )}
+      {showTextSection && (
+        <div>
+          <button onClick={handleCloseTextClick}>
+            <p>Close</p>
+          </button>
+          <NewText />
+        </div>
+      )}
+    </div>
+  );
+}
+
+export default AddText;
