+++ src/css/App.css
@@ -433,7 +433,6 @@ footer p {
   padding: 10px;
   color: black;
   margin: 0 10px;
-  width: 600px;
   background-color: rgba(255, 255, 255, 0.571);
   border-radius: 30px;
   box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.653),
@@ -442,7 +441,7 @@ footer p {
 
 .day-entries {
   text-align: center;
-  border-top: 1px solid #fc7fb1;
+  border-top: 2px solid #fc7fb1;
 }
 
 .date-picker-div {
